THomas: the profiling interface allows for using key,id couple, but it does not look like it works. Any idea ?
Change the profiling system to track payload sends/ and payload recieves. Also decreased the number of concurent sends to 1. This solved the test not being sucessful, so I had probably messed something in the multiplexing logic. I'll check it tomorrow.

The payload mismatch may be the cause of a significant part of the overhead, as reducing the number of concurent send doubled the performance.