Changed allocation of cmd to use calloc (in an attemps to avoid fragmentation with concurent tile mallocs)
More debug output to track any wrongness in the multiplexing algorithm. It actually looks ok, and the previous statemenet saying that it was related to the failed results was probably false. I just decreased the size of the test matrix and therefore got correct results, without the fact that using only one pending req having any influence. The bug is still in the wild....