Removed debug version of activate for non distributed build that was never called anyway.
There is more problems with multiple deps. The generated code doesn't do the right thing on the receiver side. The receiver needs to recompute the flags as they where computed by the original process and (or if we transmit the flags over the network) needs to compute data which are local and needs to be transferred (and datatypes and stuffs).