Split the epilog in a separate function. Add a high level interface to call the epilog function. As a result, the life of an execution context is now divided in three separate stages: the execution, the release dependencies and the epilog (which automatically call the release dependencies).
One can now save the execution context in the hook, and push it in some
other management structures. In this specific case, the hook can call
"return" directly, with any kind of negative value. Later, when the
execution of the task is comsidered completed (by the new management
system), a call to dplasma_complete_execution should be made. After
the call to dplasma_complete_execution, the execution context should
NOT be used (it is release in the epilog).