Do not use char* with DAGuE. If you do, you will be dommed for the eternity.
The reason is that fact that internally DAGuE use the least significant
bits of a ponter to store additional information. Thus, if a pointer is
not at least 2 bytes aligned, this will not work, and we will
consider thereturn of the data_of as an arena chunk (for those who
really want to dog into the details).