A simple commit with a large impact. Please read the log message carrefully.
In addition to optimizing the output code significantly, this commit introduce
exclusive input dependencies. Instead of a suite of if, where we expect the
jdf generator to enforce the uniqueness of each input, we now enforce it in the
generated C code. The most interesting addition is that now the input can
go from the most particular to the most general, hugely simplifying the
jdf.