Change the approach to compute the critical path.
Now, the critical path is computed by letting the normal scheduling happen, and keep track of the theoretical execution dates.
This improves the functionality of simulation: now the cost of a kernel can depend on the parameters of the kernel, and not only on its name.

The DOT file will output the termination date of the task, if DAGUE_SIM is enabled together with DAGUE_PROF_GRAPHER.

--HG--
branch : new_scheduling