Two weeks two lines. The most efficient code ever!!!
This was a really tricky bug to find. A race condition, difficult to
trigger, and only in the starvation case. The problem was that the
ref_count of the newly created data was originally set to zero and
increased later, once the count of all the local usages has been
done. Meanwhile, another thread might have been matched the data
for another task, and generate/schedule/execute this tasks,
prematurely decrementing the ref counter to zero, while the MPI
thread was still counting. The solution is to force the MPI thread
to get a ref count on all the data, for as long as it plays with
them. This ref count will be release at the end, once all the
local tasks have been generated. CQFD!