First step toward Control Gather. The choice between counter and mask to manage the dependencies is now per task, and can be changed at daguepp time.
Users can choose if a task readiness is handled by a bitmask or a counter, using:
  - the DAGUE_SCHED_DEPS_MASK ccmake option (expert mode) to define the default (default default is to use MASK)
  - the deps_mask property of tasks
  - the deps_count property of tasks
The compiler tries to grant the default, overloaded by the property (in the order mask better than count).
If it cannot grant an explicitely (per property) request, it will issue a warning.

Reasons for the impossibility to not grant the request:
 - the number of flows is bigger than the number of available bits in the mask allows
 - soon: the task uses control gather technique, and the number of dependencies must be counted.

The MASK technique provides much more debugging checks than the counting technique.
The counting technique will do its best to signal over-triggered dependencies, but contrarily to the mask technique, it will not point to a single over-triggered flow.

(This patch also adds a paranoid assert on the number o local tasks)