Remove some subtelty from arenas without changing the runtime behavior.
Explicitly put a dague_list_item_t at the beginning of the chunks (instead of casting them as list_item_t or chunk depending on the context).
Note that on dancer, this does not change the size of the structure at all, since this enabled to remove some fields from the chunk structure, and the size of the (removed) _internal_chunk_t was similar to the size of the actual dague_chunk_t.

Hopefuly, this makes debugging of arena reference counts easier for JDF writers.