Stronger, better, faster.
Make sure we are making a difference between the local completion flags
and the output_mask that goes on the wire. The first one contains the
bitmaks of dep_index released, while the later is the bitmask of all
the datatype_index that are made available by the remote task
completion.

--HG--
branch : qr_lu