Fix the allocation of dependencies when the ordering of the parameters is not the ordering of the locals.
Assume we have
GEMM(m, n, k)
  k = 0..MT-3
  m = k+1 .. MT-1
  n = m+1 .. NT-1

Because m and n depend on k, we want to discover the execution space with the nesting
for(k)
  for(m)
    for(n)

But then, in the dependencies array, we store by user-ordering dimensions:
deps starts with m, then deps->u.next[m] is n, then deps->u.next[m]->u.next[n] defines k

But for all possible values of k, different values of m and n are possible, so we can't just assume that once for this m and n we have generated the k dimension we're done: we need to iterate on the other values of k to see if they generate new values of m and n. Because m and n depend on k, not the opposite.

--HG--
branch : all-about-data