Lists cleanup.
No more ulifo and ufifo.
In an effort to consolidate the usage of our lists fall back to use the
list_nolock instead of all the other variants.

Never implement a function twice if there is no tangible benefit
(basically if the lock protected section span over the entire nolock
version of the function).