Addresses issue #123
The test STDC_VERSION__ >= 201112L && !__STDC_NO_ATOMICS__ is not sufficient: there are cases (e.g. gcc-6.3.1 with flag -fopenmp) where the compiler version is bigger than 201112, and __STDC_NO_ATOMICS__ is not defined, but yet, including stdatomic.h makes fail. This is due to the fact that Open MP is not defined for memory models above C99.

This patch propagates the PARSEC_HAVE_C11_ATOMICS to the PaRSEC config headers