Revert ccdacb36 and fb9189d2.
Protecting the aba_key when we are using 128 bits atomics sounded like a good idea
but unfortunately it backfired. On an external usage of PaRSEC the internal defines
of the supported atomic infrastructure are not visible, so we should not used them
to protect structure definitions (or bad things will happen as the structure will
have a different size inside and outside PaRSEC).

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>