Base infrastructure to enable PAPI SDE.
papi_sde_interface.c and papi_sde_interface.h are stubs provided by
PAPI SDE that allow to compile and run without PAPI SDE and without
configure options. They are compiled into a dynamic library that is
added after any other library and they provide the necessary weak
symbols. If libpapi.so is added (through PINS or another mechanism)
before that dynamic library, and that version of papi provides the
hooks, those are used instead.

Change the parser in parsec_pins_papi_events_new with two goals:
  - Allow event names to include ':' in their string. PAPI uses
    ':' as a namespace separator, so using ':' to separate multiple
    events was a problem. This replaces it with '|' to separate
    multiple events
  - Allow event names to start with 'C', 'S' or 'F'. There was no
    way in the current code to have an event starting with one of
    these letters. Change the keyword to #C #S and #F to denote
    Core, Socket or Frequency.

Move where the pins_papi_event MCA parameter string is read later.

If we want to read the PAPI Software Defined Events that we expose,
then this string must be read after the modules that expose them
loaded. Moving it here allows the schedulers to define them.

Define functions to know the length of the dequeue and the bounded buffers.

Insert PAPI SDE software event defined to know the lenght of
scheduling queues for the LFQ scheduler (other schedulers still
to be done).

Signed-off-by: Thomas Herault <herault@icl.utk.edu>