Infrastructure to enable PAPI SDE.
This is the base infrastructure to provide internal counters of PaRSEC to external tools and applications using the PAPI Software Defined Event interface.

The approach taken by PAPI SDE is to provide weak symbols for the PAPI SDE calls (e.g. register_counter), that do nothing by default, but when linked with a PAPI-SDE enabled library, expose the counters through PAPI. I integrated that by providing the interface as a dynamic library that is added to the link line in last position, so if one links with PAPI, we can use the SDE counters we define as MCA parameters for PINS and read them inside the profiling system of PaRSEC.

Remove thread-unsafe length function

Remove dupplicate declaration due to merge error

Rename length function to clearly specify that it only approximates the number of occupied cells in the bounded buffer

Add comments to specify why there are two cases in sde_register

Move PINS mca registration to original place; keep the parameter parsing to after all MCA modules have initialized; add a comment to explain why

Signed-off-by: Thomas Herault <herault@icl.utk.edu>