Fix parsec_pins_papi_events_new parser to accept PAPI counter names that start with C, F or S.
The parser iterates over each ':', looks at the character that follows the ':', and if it is C, F or S, interprets the next characters as a number or a frequency. As a consequence, counters like SSE_MEM_EXEC could not be read by the parser: even if passed as C*:F*:S*:SSE_MEM_EXEC, as all 'S' after a ':' is a socket specifier.

With this patch, once a specifier has been set for a counter, the character can be used in the counter name.