Fix the use of layouut in output flows.
It has been reported that in some cases specifying only the type of an
output flow might lead to generating incorrect code. The reason was a
shortcut in the code generation where a missing layout was considered
the same as the type. This patch addresses this issue and allows for
less code on the type/layout description.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>