Simplify the hash table API
This PR follows a discussion on the hash table API: the current
API in master requires the user to provide a hashing function
to hash the items on a variable number of bits. If the hashing
function provided does not ensure a high probability of collision
avoiding when re-hashing the same keys for different number of bits,
this translates into memory exhaustion wiht high probability.

The proposed API requires the user to provide a hashing function
to map an item key onto 64 bits. Then, all internal re-hashing from that
hash uses a universal hash function that has the desired property.

From the API point of view, this translates into a removal of the
nbits parameter for the user-defined hash function. We only use
the non-default top-level hashing function in three places in
the current code, so changes are minimal outisde parsec_hash_table.c