Refactor info mechanism as a separate entity from the object / class system, so we can decide what objects are extended with infos. Infos still come with two separate high-level objects: parsec_info_t is a dictionary of keys (strings) that provides a unique index, and parsec_info_object_array_t stores values in dynamic arrays.
Enable CUBLAS-v2 and other stateful device management using Info keys.

Use cublas_v2.h as a demonstration in the nvlink test:
the wrapper creates one cublasHandle_t per exec stream
and per device, and exposes them as info keys under
CUBLAS::HANDLE, then the body of the JDF uses the
actual version 2 interface of CUBLAS. Note that the
API does not simply add the handle parameter, it also
changes other parameters of GEMM (how transpose and
alpha beta are passed, notably).

The same info system can be used to pass a cuDnSolveHandle,
and potentially pointers to temporary GPU memory.

To add some context: some libraries (e.g. CUBLAS v2, and
cu*Solve) now use a 'handle' attached to a specific stream
of a specific device instead of relying on the current
cudaStream to submit work to the GPUs. Creating these handles
and attaching them to streams for each kernel is too costly.
The info key approach aims to be a portable way (i.e. that
is not fit only for CUBLAS and cuDnSolve) to address these
issues.

Infos are key-value stores: keys identify infos,
giving them a single unique ID that is used to find the values
in arrays.

Addressed PR comments and questions; added comments where code
was unclear; added or update copyright in all modified files