Hotfix: abusive test for NULL flow in cuda device.
PR #330 introduced a check to ignore any input flow that
is NULL. This is mostly correct, except for this case.

The test for NULL flow is already in parsec_gpu_data_reserve_device_space
at line 814, and doing it at line 801 was wrong, as this skipped initializing
temp_loc[i] to NULL for that flow. As temp_loc is an array on the stack of
the function, this makes possible to detect a cycle and announce lack of
memory by mistake.