Fixes one problem, but cannot reproduce Issue 225 yet.
When the same task reads two flows that point to the same data copy,
the current code would assert:
  - first, the data copies are allocated, and the code already managed
    this case properly
  - but then, flows copies are initiated one after the other, and the
    code would find the second flow (pointing to the same data copy
    that was already put in transfer) already in transfer, triggering
    an assert.

In this version, if a data copy is already in transfer inside push_in,
it is assumed that this transfer is the one we need, and no assert
triggers, and the data copy is managed as if it initiated the transfer
(so it's not marked ready yet either).

This makes the test in tests/interface/ptg/cuda/stress* successfully
complete, so this does not help reproduce Issue #225. More information
on the issue is requested.