All generated taskpools are now derived from taskpool_t.
Important: the internal taskpool is derived from the the external
taskpool, so the highest taskpool in the hierarchy is not the JDF-named
taskpool but the internal one. Thus, any OBJ_NEW for such a taskpool
should be done on the internal taskpool object.

Added a new option no_taskpool_instance to minimize the impact on users in
straighforward cases. This option prevents the PTG compiler from generating
the taskclass instance allowing the user to define it's own.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>