Addresses Issue #239 by replacing simple hash table with parallel one.
The hash table implementation used in tree_dist.[ch] is not thread-safe:
it decorelates the node identifier allocation from accessing the nodes
themselves. As nodes can be rehashed any time by another thread, it's
possible that tree->nballocation changes between the call to lookup_node
(that returns an nid) and the derference tree->nodes[nid]. If that variable
changed, then the right node is not in tree->nodes[nid] any more.

Instead of using yet another implementation, use the parsec_hash_table_t
that is thread safe and tested on many other cases to store nodes of the
tree.

Also cleanup the tree after the taskpool is done.