Correctly generate the parsec_options.h
Due to the fact that we need this file during the configure to determine
how to handle the 128 bits atomics, we need to regenerate it again later
to account for later updates to the exported variables. Fix the logic
and add comments to remember why the things are done the way they are.

Remove all source and build tree dependencies from ptgpp.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>