Cygwin and WLS [checked].
A lot of changes to add support for Cygwin and Windows Linux subsystem
INPUT being a default structure on Windows, we need to use a namespace for ours. Prefix all DTD flags with PARSEC_.
Same story for IN and OUT macros.
Smoothly handle the \r\n on Windows, to allow the generated files to be
read by the native compiler.
Cleanup a lot of warnings.
Don't use mmap on Windows
 - prior solution was brutal, it disabled the profiling. A nicer solution is to disable the mmap use in the profiling, but allow the profiling to be in use.
PRovide support for random generator and for environment on Windows.
Fix TAU discovery.
Add missing files to the project.
Disable the aggregator tool on Windows (due to missing support for mmap).

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>