Improve the CUDA search code.
First check with the current capability, and only then with one of the
capability from the static array. At the end check for the function without any
extension. Also, if the capability is lesser than the first value in the array
check it too, and if it is larger after checking it fall back to tne entire
array.

Be more forgiving on the identification of the supported devices.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>