Create a profiling test for the ASYNC and AGAIN return codes in PTG.
Remove abusive shortcut that would disregard events with ID == 0 ending on a thread different from the one they started on

Add command-line parsing for async.jdf test

Start automatizing the tests.

Adapt test to new interface for two_dim_block_cyclic_init

Add arguments to profile2h5:
  - allow to enable/disable multiprocessing and process tracking
  - allow to specify output filename

When the output filename is specified, the script does not try to group input files by execution, allowing to bypass the checks that those belong to the same run.

Working tests for the profiling (for async and again specifically).