Backend file should stay open until the end.
Do not allow the profiling to close the backend file in the middle of the
execution, not even when mmap fails. Instead, mark the file as non extendable
(such that existing profiling buffers will be written but no new buffers
can be allocated).
Add some documentation and change the name of few functions.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>