Yarrow's ifx compiler does not support Wall and Wextra and mcx16.
This commit disables Wall and Wextra for all Fortran Intel compilers,
and no Fortran file uses the atomics directly, so none of them
need mcx16. This removes the flag for all Fortran files.

It would be better to use target_set_properties and source_set_properties
and get rid of this global compiler_options that affects all compilers
even though we tested the flags only against the C compiler.