There was a subtle race condition in dynamic JDFs like project.jdf
In those, tp->nb_tasks is initialized to UNDETERMINED_NB_TASKS
and the generated code does not update nb_tasks if it is
UNDETERMINED or 0. The user code is supposed to set tp->nb_tasks = 0
in a task that is the last one to run.

However, the last task body can run while other threads are still
finishing to process the release_deps function of ancestor tasks.
Then, they run the release_task function on their tasks.

If the hook has set tp->nb_tasks to 0, one of them can detect it
and call atomic_swap(tp->nb_tasks, 0, RESERVED) to mark that the
runtime considers this taskpool is done from the user perspective
(which is true, all bodies have run).

So, if another ancestor comes after that and finally executes its
release_task, the generated code did not test that case: RESERVED
is not UNDETERMINED or 0, and it would decrement nb_tasks like a
normal behavior.

This would eventually make a nb_tasks that goes to down to
-nb_cores+2 in the worst case, and trigger asserts during the
taskpool cleanup.

The solution here says to the generated release_task that RESERVED
is also a number that it cannot touch.