Add DPC++ to the loop...
  - Add multiple CMake logic files and commands
  - jdf2c.c now generates dpcpp output files when needed
  - make DEV_DPCPP be an alias to DEV_LEVEL_ZERO
  - Command Lists for I/O (streams of id 0 and 1) are still immediate
  - Command Lists for computations (streams of id >= 2) are now normal lists connected to a queue
    that queue exists as a compute level-zero queue and as a DPC++ queue
  - Missing compilation logic to compile generated dpc++ code and link it with the target binary

Risk: it is unclear that the user can still push orders / events in the command list, after it is closed,
and it is necessary to close it to force the orders to be pushed on the queue. I might need to create a
new command list after each close, and attach the command list to the event for garbage collection.

Adapt findlevel-zero.cmake to support systems where pkg-config is broken