Use explicit memory ordering for atomic locks
- atomic_lock uses memory_order_acquire:
  A load operation with this memory order performs the acquire
  operation on the affected memory location: no reads or writes
  in the current thread can be reordered before this load. All
  writes in other threads that release the same atomic variable
  are visible in the current thread
- atomic_unlock uses memory_order_release:
  A store operation with this memory order performs the release
  operation: no reads or writes in the current thread can be
  reordered after this store. All writes in the current thread
  are visible in other threads that acquire the same atomic variable
  and writes that carry a dependency into the atomic variable become
  visible in other threads that consume the same atomic

See https://en.cppreference.com/w/c/atomic/memory_order#Release-Acquire_ordering

Note that the acquire-release semantic is not limited to either load
or store operations but applies to both. On x86-64, the unlock
can be implemented as a simple store instead of an xchg because x86
is strongly ordered anyway and single-byte stores are atomic.
According to https://www.felixcloutier.com/x86/xchg, the xchg instruction
activates the processors locking protocol, which adds unnecessary overhead.
On Arm, the compiler might use a weaker synchronization instruction
than with sequential consistency.

Atomic spinlock is a classic example for acquire-release semantic, so
we should use it :)

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>