Allow sender to send data of any size.
The sender can send less data than expected by the receiver. If we
select the communication protocol based on how much data the receiver
expects, the sender and receiver could diverge on the protocol to
be used (mainly visible when the sent message is below eager, while
the receiver expects more data. So, force the sender to pack the
amount of data per dep, and force the receiver to abide by this amount.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>