Add support for CUDA new data copies.
Signed-off-by: Thomas Herault <herault@icl.utk.edu>

When discovering nvcc, CMake will check that nvcc accepts the same compile options as the parsec target, because the testcase that uses a cu file depends on the target parsec... So, we need to deactivate the sanitizing options (that are not supported by nvcc) when using a compiler that is not the C compiler.

nvcc of cuda 9.2.88 does not know how to compile our dtd_test_new_tiles_cuda_kernels.cu file in Release mode.

Port dtd_test_new_tile to new PaRSEC data collection API (remove deprecated warnings)

DTD (maybe among other users of the hash tables) has memory
leaks in the main structures of the hash tables it creates.

parsec_hash_table_t needs a special init that PARSEC_OBJECT_NEW
or PARSEC_OBJECT_CONSTRUCT could not provide. This is
parsec_hash_table_init.

So, it was assumed that by symmetry, the user would call
parsec_hash_table_fini on it.

However, a parsec_hash_table_t is also a parsec_object_t,
so calling PARSEC_OBJECT_DESTRUCT() or PARSEC_OBJ_RELEASE()
with a usage count that falls to 0 should release all memory
allocated by the object.

This patch proposes to solve the solution by making sure
that parsec_hash_table_fini can be called multiple times on
the same hash table without segfaulting (so if users call
explicitly parsec_hash_table_fini, this doesn't fail), while
at the same time automatically calling parsec_hash_table_fini
on the object if the user calls PARSEC_OBJECT_DESTRUCT or
PARSEC_OBJECT_RELEASE with the right condition.

rework DTD NEW tile example to show communication

Remove warning about deprecated function usage

Make sure the GEMM test is not used for performance measuring

Remove debug code that is never used

PARSEC_PUSHOUT is only defined if we have a device that uses PUSHOUT

Fix bug in the logic for handling NEW data in DTD at the GPU level:
the GPU does not need to pull the data from another place iff
   - it is NEW (no repo entry)
   - AND the current data version is 0, denoting that no previous
     task has written something in the data...

Rename functions to simplify debugging, add some
prints to simplify debugging, AND solve a bug (PARSEC_PUSHOUT
missing on the output of the last task if it runs on a GPU).

Factor parsec_get_data_copy and arena_device_private_allocate

Add the test_new_tile as part of the validation runs

Fix wrong arguments for profiling (appeared during a merge?)

If using CUDA in DTD, device_cuda.h needs to be included so that the user code can write the CUDA host code