Force the same alignement for the external and internal lifo.
The real issue was with the external version of the parsec_lifo_t. In C a structure inherits the most restrictive alignment of its members. Thus, on architectures with support for 128 bits integers the internal structure will be aligned by 16 bytes. Unfortunately, the external structure, aka. parsec_lifo_opaque_s, will inherit the alignment of int64_t. Thus, in most cases where the lifo is allocated via OBJ_NEW, things will just work because malloc will return a correctly aligned memory segment. But in the test the lifo were static objects, and thus their alignment was decided at compile time, based on the 8 bytes alignment of the external lifo. The fix is to add to the external struct a union containing an __int128 type, which will then propagate the restrictive alignment to the external struct.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>