Provide a temporary fix for the flag detection.
This is not yet perfect, as the output is misleading because it includes
the name of the variable to define instead of being the flag itself.

However, I don't think this works as intended as on my machine despite
having a succesful detection for -fdiagnostics-color it is not added to
the compilation flags.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>